Optimized water buffer usage

+++ shaders/lib/Settings.glsl
@@ -12,7 +12,7 @@ const float wetnessHalflife          = 200.0;
 const float drynessHalflife          = 40.0;
 
 /*
-const int colortex0Format = RG32F;
+const int colortex0Format = RG16;
 const int colortex1Format = RGB16;
 const int colortex2Format = RGB16;
 const int colortex3Format = RGB32F;
